+++ app/src/main/java/com/example/nasimuzzaman/roostpad/updateUser/UpdateUserResponse.java
@@ -13,7 +13,7 @@ public class UpdateUserResponse extends BaseResponse {
     @SerializedName("email") private String email;
     @SerializedName("role") private String role;
     @SerializedName("designation") private String designation;
-    @SerializedName("holiday") private int holiday;
+    @SerializedName("holiday") private double holiday;
     @SerializedName("gender") private String gender;
     @SerializedName("token") private String token;
     @SerializedName("name") private String name;
@@ -34,7 +34,7 @@ public class UpdateUserResponse extends BaseResponse {
         return designation;
     }
 
-    public int getHoliday() {
+    public double getHoliday() {
         return holiday;
     }
